basecli tx handles json input
