+++ client/commands/txs/helpers.go
@@ -4,8 +4,6 @@ import (
 	"bufio"
 	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"os"
 	"strings"
 
@@ -34,7 +32,7 @@ type Validatable interface {
 // GetSigner returns the pub key that will sign the tx
 // returns empty key if no name provided
 func GetSigner() crypto.PubKey {
-	name := viper.GetString(NameFlag)
+	name := viper.GetString(FlagName)
 	manager := keycmd.GetKeyManager()
 	info, _ := manager.Get(name) // error -> empty pubkey
 	return info.PubKey
@@ -53,7 +51,7 @@ func GetSignerAct() (res basecoin.Actor) {
 
 // Sign if it is Signable, otherwise, just convert it to bytes
 func Sign(tx interface{}) (packet []byte, err error) {
-	name := viper.GetString(NameFlag)
+	name := viper.GetString(FlagName)
 	manager := keycmd.GetKeyManager()
 
 	if sign, ok := tx.(keys.Signable); ok {
@@ -90,32 +88,6 @@ func SignAndPostTx(tx Validatable) (*ctypes.ResultBroadcastTxCommit, error) {
 	return node.BroadcastTxCommit(packet)
 }
 
-// LoadJSON will read a json file from disk if --input is passed in
-// template is a pointer to a struct that can hold the expected data (&MyTx{})
-//
-// If not data is provided, returns (false, nil)
-// If data is provided and passes, returns (true, nil)
-// If data is provided but not parsable, returns (true, err)
-func LoadJSON(template interface{}) (bool, error) {
-	input := viper.GetString(InputFlag)
-	if input == "" {
-		return false, nil
-	}
-
-	// load the input
-	raw, err := readInput(input)
-	if err != nil {
-		return true, err
-	}
-
-	// parse the input
-	err = json.Unmarshal(raw, template)
-	if err != nil {
-		return true, err
-	}
-	return true, nil
-}
-
 // OutputTx prints the tx result to stdout
 // TODO: something other than raw json?
 func OutputTx(res *ctypes.ResultBroadcastTxCommit) error {
@@ -138,25 +110,6 @@ func signTx(manager keys.Manager, tx keys.Signable, name string) ([]byte, error)
 		return nil, err
 	}
 	return tx.TxBytes()
-}
-
-func readInput(file string) ([]byte, error) {
-	var reader io.Reader
-	// get the input stream
-	if file == "-" {
-		reader = os.Stdin
-	} else {
-		f, err := os.Open(file)
-		if err != nil {
-			return nil, err
-		}
-		defer f.Close()
-		reader = f
-	}
-
-	// and read it all!
-	data, err := ioutil.ReadAll(reader)
-	return data, errors.WithStack(err)
 }
 
 // if we read from non-tty, we just need to init the buffer reader once,
